+++ test/PriceFeed.gas.test.ts
@@ -1,6 +1,6 @@
 import { parseEther } from "ethers/lib/utils"
 import { ethers, waffle } from "hardhat"
-import { BandPriceFeed, ChainlinkPriceFeed, TestAggregatorV3, TestPriceFeed, TestStdReference } from "../typechain"
+import { BandPriceFeed, ChainlinkPriceFeedV2, TestAggregatorV3, TestPriceFeed, TestStdReference } from "../typechain"
 
 const twapInterval = 900
 interface PriceFeedFixture {
@@ -9,7 +9,7 @@ interface PriceFeedFixture {
     baseAsset: string
 
     // chainlinik
-    chainlinkPriceFeed: ChainlinkPriceFeed
+    chainlinkPriceFeed: ChainlinkPriceFeedV2
     aggregator: TestAggregatorV3
 }
 
@@ -30,11 +30,11 @@ async function priceFeedFixture(): Promise<PriceFeedFixture> {
     const testAggregatorFactory = await ethers.getContractFactory("TestAggregatorV3")
     const testAggregator = await testAggregatorFactory.deploy()
 
-    const chainlinkPriceFeedFactory = await ethers.getContractFactory("ChainlinkPriceFeed")
+    const chainlinkPriceFeedFactory = await ethers.getContractFactory("ChainlinkPriceFeedV2")
     const chainlinkPriceFeed = (await chainlinkPriceFeedFactory.deploy(
         testAggregator.address,
         twapInterval,
-    )) as ChainlinkPriceFeed
+    )) as ChainlinkPriceFeedV2
 
     return { bandPriceFeed, bandReference: testStdReference, baseAsset, chainlinkPriceFeed, aggregator: testAggregator }
 }
@@ -44,7 +44,7 @@ describe.skip("Price feed gas test", () => {
     const loadFixture: ReturnType<typeof waffle.createFixtureLoader> = waffle.createFixtureLoader([admin])
     let bandPriceFeed: BandPriceFeed
     let bandReference: TestStdReference
-    let chainlinkPriceFeed: ChainlinkPriceFeed
+    let chainlinkPriceFeed: ChainlinkPriceFeedV2
     let aggregator: TestAggregatorV3
     let currentTime: number
     let testPriceFeed: TestPriceFeed